refactor

+++ src/com/gto/phoebe/Glue.java
@@ -2,9 +2,6 @@ package com.gto.phoebe;
 
 import java.awt.*;
 
-/**
- * Created by atesz17 on 3/5/2015.
- */
 public class Glue extends Actor {
 
     public Glue()    {
@@ -17,7 +14,7 @@ public class Glue extends Actor {
 
     @Override
     public void activateEffectOn(Robot robot) {
-        int newSpeed = (int)Math.round(robot.getSpeed()/2.0);
+        int newSpeed = Math.round(robot.getSpeed() / 2.0f);
         robot.setSpeed(newSpeed);
     }
 }
